- fixed network graphic - debuged evaluation tables - changed cache settings in template engine - some speed hacks - changed int angles for peer positions in network graphic to double angles
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@8124 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/net/yacy/kelondro/blob/MapHeap.java
@@ -194,7 +194,7 @@ public class MapHeap implements Map<byte[], Map<String, String>> {
             if (this.cache != null) this.cache.remove(key);
 
             // remove from file
-            this.blob.delete(key);
+            if (this.blob != null) this.blob.delete(key);
         }
     }
